Move guts of showcase-detail to seperate component as can't have regular page + staticquery in same file
